Refactor: ExpenseDetail, AddExpense 관련된 BindingAdapter 이동
