+++ app/src/main/java/com/kappzzang/jeongsan/util/BindingAdapter.kt
@@ -14,58 +14,4 @@ import kotlinx.coroutines.flow.StateFlow
 
 object BindingAdapter {
 
-    @JvmStatic
-    @BindingAdapter("expenseImageUrl")
-    fun loadImage(view: ImageView, url: String?) {
-        if (!url.isNullOrEmpty()) {
-            Glide.with(view.context)
-                .load(url)
-                .into(view)
-        } else {
-            // 기본 이미지 또는 처리 로직 추가 가능
-        }
-    }
-
-    @BindingAdapter("stringText")
-    @JvmStatic
-    fun setTextValue(view: EditText, value: String?) {
-        view.setText(value ?: "")
-    }
-
-    @InverseBindingAdapter(attribute = "stringText")
-    @JvmStatic
-    fun getTextValue(view: EditText): String? {
-        val parsedString = view.text.toString()
-        if (parsedString.isBlank()) {
-            return null
-        }
-        return parsedString
-    }
-
-    @BindingAdapter("integerText")
-    @JvmStatic
-    fun setIntegerValue(view: EditText, value: Int?) {
-        view.setText(value?.toString() ?: "")
-    }
-
-    @InverseBindingAdapter(attribute = "integerText")
-    @JvmStatic
-    fun getIntegerValue(view: EditText): Int? = view.text.toString().toIntOrNull()
-
-    @BindingAdapter("integerTextAttrChanged")
-    @JvmStatic
-    fun setIntegerTextListener(view: EditText, listener: InverseBindingListener) {
-        view.addTextChangedListener {
-            listener.onChange()
-        }
-    }
-
-    @BindingAdapter("stringTextAttrChanged")
-    @JvmStatic
-    fun setStringTextListener(view: EditText, listener: InverseBindingListener) {
-        view.addTextChangedListener {
-            listener.onChange()
-        }
-    }
-
 }
