+++ ui/addexpense/build.gradle.kts
@@ -6,5 +6,4 @@ dependencies {
     implementation(project(":domain:ocr"))
     implementation(project(":ui:expensedetail"))
     implementation(project(":domain:expense"))
-    implementation(project(":app"))
 }
