+++ ui/expensedetail/build.gradle.kts
@@ -3,5 +3,7 @@ android {
 }
 
 dependencies {
+    implementation("com.github.bumptech.glide:glide:4.14.2")
+    annotationProcessor("com.github.bumptech.glide:compiler:4.14.2")
     implementation(project(":domain:expense"))
 }
